refactor: group governor hub tx
